tweaked damage and speed values on tools

+++ src/main/java/me/FenrisFox86/ancientcores/common/items/core/CoreShovel.java
@@ -23,7 +23,7 @@ public class CoreShovel extends ShovelItem implements ICoreItem {
 
     public final ICoreType core;
 
-    public CoreShovel(ICoreType core, int attackDamageIn, float attackSpeedIn) {
+    public CoreShovel(ICoreType core, float attackDamageIn, float attackSpeedIn) {
         super(
                 core.getItemTier(),
                 attackDamageIn,
